Add files via upload
Apply Apriori Algorithm to the Market Dataset.
Apriori is an algorithm for frequent item set mining and association rule learning over relational databases. It proceeds by identifying the frequent individual items in the database and extending them to larger and larger item sets as long as those item sets appear sufficiently often in the database. This algorithm uses two steps “join” and “prune” to reduce the search space. It is an iterative approach to discovering the most frequent item-sets.